micro service question update and controller folder location correction
