+++ src/main/java/com/riwi/filtroSpringBoot/api/controllers/QuestionsController.java
@@ -1,7 +1,7 @@
-package com.riwi.filtroSpringBoot.controllers;
+package com.riwi.filtroSpringBoot.api.controllers;
 
-import org.springframework.data.domain.Page;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,9 +13,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.riwi.filtroSpringBoot.api.dto.request.UserRequest;
-import com.riwi.filtroSpringBoot.api.dto.response.UserResponse.UserResponse;
-import com.riwi.filtroSpringBoot.infraestructure.abstract_services.IUserEntityService;
+import com.riwi.filtroSpringBoot.api.dto.request.QuestionsRequest;
+import com.riwi.filtroSpringBoot.api.dto.response.QuestionOptionsResponse.QuestionsResponse;
+import com.riwi.filtroSpringBoot.infraestructure.abstract_services.IQuestionsService;
 import com.riwi.filtroSpringBoot.util.enums.SortType;
 import com.riwi.filtroSpringBoot.util.exceptions.ResourceNotFoundException;
 
@@ -27,14 +27,16 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.AllArgsConstructor;
 
 @RestController
-@RequestMapping(path = "/users")
+@RequestMapping(path = "/questions")
 @AllArgsConstructor
-@Tag(name = "User Entity Controller")
-public class UserEntityController {
+@Tag(name = "Questions Entity Controller")
+
+public class QuestionsController {
+
 
 
     @Autowired
-    private final IUserEntityService iUserEntityService;
+    private final IQuestionsService iQuestionsService;
 
     /*----------------------------
      * GET ALL
@@ -42,7 +44,7 @@ public class UserEntityController {
      */
 
      @Operation(
-        summary = "Displays all Users",
+        summary = "Displays all Questions",
         description = "Displays the Users in a list, it is configured to display 10 items per page."
     )
     @ApiResponses(value = {
@@ -53,12 +55,12 @@ public class UserEntityController {
         @ApiResponse(responseCode = "500", description = "INTERNAL SERVER ERROR")
     })
      @GetMapping
-    public ResponseEntity<Page<UserResponse>> getAll(
+    public ResponseEntity<Page<QuestionsResponse>> getAll(
             @Parameter(description = "Page number (default: 1)", example = "1") // SWAGGER
             @RequestParam(defaultValue = "1") int page,
             @Parameter(description = "Number of items per page (default: 10)", example = "10") // SWAGGER
             @RequestParam(defaultValue = "10") int size) {
-        return ResponseEntity.ok(this.iUserEntityService.getAll(page - 1, size, SortType.NONE));
+        return ResponseEntity.ok(this.iQuestionsService.getAll(page - 1, size, SortType.NONE));
     }
 
 
@@ -68,8 +70,8 @@ public class UserEntityController {
      */
 
     @Operation(
-        summary = "Displays one user by id",
-        description = "Shows the user by the ID sent or requested by path,value cannot be less than 1."
+        summary = "Displays one question by id",
+        description = "Shows the question by the ID sent or requested by path,value cannot be less than 1."
     ) //SWAGGER
     @ApiResponses(value = {
         @ApiResponse(responseCode = "201", description = "SUCCESSFUL"),
@@ -78,26 +80,26 @@ public class UserEntityController {
         @ApiResponse(responseCode = "403", description = "FORBIDDEN ACCESS"),
         @ApiResponse(responseCode = "500", description = "INTERNAL SERVER ERROR")
     })
-    @GetMapping(path = "/{user_id}")
-    public ResponseEntity<UserResponse> getById(
-        @Parameter(description = "User ID",example = "1") // SWAGGER
-        @PathVariable Integer user_id) {
+    @GetMapping(path = "/{question_id}")
+    public ResponseEntity<QuestionsResponse> getById(
+        @Parameter(description = "Question ID",example = "1") // SWAGGER
+        @PathVariable Integer question_id) {
 
-        UserResponse user = this.iUserEntityService.getById(user_id);
-        if (user == null) {
-            throw new ResourceNotFoundException("User not found");
+        QuestionsResponse question = this.iQuestionsService.getById(question_id);
+        if (question == null) {
+            throw new ResourceNotFoundException("Question not found");
         }
-        return ResponseEntity.ok(user);
+        return ResponseEntity.ok(question);
     }
 
      /*--------------------
-     * CREATE USER
+     * CREATE QUESTION
      * -------------------
      */
 
     @Operation(
-        summary = "creates a new user",
-        description = "create a new user by entering the required data"
+        summary = "creates a new question",
+        description = "create a new question by entering the required data"
     ) //SWAGGER
     @ApiResponses(value = {
         @ApiResponse(responseCode = "201", description = "SUCCESSFUL"),
@@ -107,18 +109,18 @@ public class UserEntityController {
         @ApiResponse(responseCode = "500", description = "INTERNAL SERVER ERROR")
     })
     @PostMapping(path = "/create")
-    public ResponseEntity<UserResponse> create(@Validated @RequestBody UserRequest request) {
-        return ResponseEntity.ok(this.iUserEntityService.create(request));
+    public ResponseEntity<QuestionsResponse> create(@Validated @RequestBody QuestionsRequest request) {
+        return ResponseEntity.ok(this.iQuestionsService.create(request));
     }
 
      /*----------------------
-     * UPDATE USER
+     * UPDATE QUESTION
      * ---------------------
      */
 
      @Operation(
-        summary = "update  user by ID",
-        description = "updates a previously created user and the ID and the new modified parameters must be sent through the Path, value cannot be less than 1"
+        summary = "update  question by ID",
+        description = "updates a previously created question and the ID and the new modified parameters must be sent through the Path, value cannot be less than 1"
     ) //SWAGGER
     @ApiResponses(value = {
         @ApiResponse(responseCode = "201", description = "SUCCESSFUL"),
@@ -127,13 +129,12 @@ public class UserEntityController {
         @ApiResponse(responseCode = "403", description = "FORBIDDEN ACCESS"),
         @ApiResponse(responseCode = "500", description = "INTERNAL SERVER ERROR")
     })
-    @PutMapping(path = "/{user_id}")
-    public ResponseEntity<UserResponse> update(@Validated @RequestBody UserRequest request, 
+    @PutMapping(path = "/{question_id}")
+    public ResponseEntity<QuestionsResponse> update(@Validated @RequestBody QuestionsRequest request, 
         @Parameter(description = "User ID",example = "1") // SWAGGER
-        @PathVariable Integer user_id) {
-        return ResponseEntity.ok(this.iUserEntityService.update(request, user_id));
+        @PathVariable Integer question_id) {
+        return ResponseEntity.ok(this.iQuestionsService.update(request, question_id));
     }
-
